Move edit components from design/config to design/FormManager/FormEdit, so they are closer to the place they are used.
