+++ packages/design/src/FormManager/FormEdit/components/index.ts
@@ -3,7 +3,7 @@ import FormSummaryEdit from './FormSummaryEdit';
 import InputPatternEdit from './InputPatternEdit';
 import ParagraphPatternEdit from './ParagraphPatternEdit';
 import SubmissionConfirmationEdit from './SubmissionConfirmationEdit';
-import { type EditComponentForPattern } from '../../FormManager/FormEdit/types';
+import { type EditComponentForPattern } from '../types';
 
 export const defaultPatternEditComponents: EditComponentForPattern = {
   paragraph: ParagraphPatternEdit,
