+++ packages/design/src/config/index.ts
@@ -1,2 +1,1 @@
-export * from './edit';
 export * from './view';
