+++ packages/design/src/index.ts
@@ -4,4 +4,5 @@ export { default as AvailableFormList } from './AvailableFormList';
 export { default as Form } from './Form';
 export { default as FormRouter } from './FormRouter';
 export { default as FormManager } from './FormManager';
+export { defaultPatternEditComponents } from './FormManager/FormEdit/components';
 export * from './config';
